requirements.txt
python-telegram-bot[webhooks]==20.7
supabase==2.3.0
pytz==2024.1
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
